+++ src/seedu/addressbook/data/person/Contact.java
@@ -1,0 +1,44 @@
+package seedu.addressbook.data.person;
+
+/**
+ * Parent class for contact fields (Address, Email, Phone)
+ */
+
+public class Contact {
+
+    private String value;
+    private boolean isPrivate;
+
+    public Contact(boolean isPrivate) {
+        this.isPrivate = isPrivate;
+    }
+
+    public Contact() {
+    }
+
+    public void setValue(String value) {
+        this.value = value;
+    }
+
+    @Override
+    public String toString() {
+        return value;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof Address // instanceof handles nulls
+                && this.value.equals(((Address) other).value)); // state check
+    }
+
+    @Override
+    public int hashCode() {
+        return value.hashCode();
+    }
+
+    public boolean isPrivate() {
+        return isPrivate;
+    }
+
+}
